fix: remove sharedPortalArea context
