+++ src/hooks/index.ts
@@ -17,10 +17,14 @@ export * from './useNullableState'
 export * from './usePrevious'
 export * from './useRemoveSnackbar'
 export * from './useRunIfMounted'
+export * from './useSharedPortalAreaInsets'
+export * from './useSharedPortalAreaSize'
 export * from './useSnackbarSettings'
 export * from './useSnackbarsToShow'
 export * from './useSnackbarWasPresented'
 export * from './useTheme'
 export * from './useToggle'
 export * from './useToggleWithHaptics'
+export * from './useUpdateSharedPortalAreaInsets'
+export * from './useUpdateSharedPortalSafeAreaInsets'
 export * from './useWillKeyboardBeShown'
