+++ src/hooks/useSharedPortalAreaInsets.ts
@@ -1,0 +1,5 @@
+import useSharedPortalAreaStore from './useSharedPortalAreaStore'
+
+export const useSharedPortalAreaInsets = () => useSharedPortalAreaStore((state) => state.insets)
+
+export default useSharedPortalAreaInsets
