+++ src/hooks/useSharedPortalAreaSize.ts
@@ -1,0 +1,5 @@
+import useSharedPortalAreaStore from './useSharedPortalAreaStore'
+
+export const useSharedPortalAreaSize = () => useSharedPortalAreaStore((state) => state.size)
+
+export default useSharedPortalAreaSize
